Use C++ style file I/O
It is easier to work with C++ file I/O than C. Additionally, there was a problem in the previous version whereby each traversal opened the same file. Instead, we use some OOP to make it so that it only opens it once (in the constructor) and closes it when done (in the destructor).